+++ src/widgets/header/hooks/useSearchInputDebounce.js
@@ -1,6 +1,6 @@
 import { useEffect, useState } from 'react';
 
-export default function useDebounce(value, delay) {
+export default function useSearchInputDebounce(value, delay) {
   const [debouncedValue, setDebouncedValue] = useState(value);
 
   useEffect(() => {
