+++ src/widgets/header/ui/search-bar/SearchBar.container.js
@@ -1,4 +1,4 @@
-import React, { useRef, useState } from 'react';
+import { useRef, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import SearchBarUI from './SearchBar.presenter';
