+++ src/widgets/header/ui/search-bar/SearchBar.styles.js
